refactor: extract shared rate-limit/JSON controller plumbing into a trait
parseJsonBody(), jsonResponseWithRateLimitHeaders() and rateLimitedResponse()
were byte-identical copies across the AJAX controllers (the source of the
SonarCloud new-code duplication). Move them into RateLimitedControllerTrait.
AjaxController keeps its own rateLimitedResponse() (it serialises a
CompleteResponse body) — a class method overrides the trait's.
