+++ Classes/Controller/AjaxController.php
@@ -55,6 +55,8 @@ use TYPO3\CMS\Core\Type\Bitmask\Permission;
  */
 final readonly class AjaxController
 {
+    use RateLimitedControllerTrait;
+
     /**
      * Maximum number of messages in a chat conversation.
      */
@@ -924,25 +926,6 @@ final readonly class AjaxController
         }
 
         return $response->withAddedHeader('Retry-After', (string) $result->getRetryAfter());
-    }
-
-    /**
-     * Create JSON response with rate limit headers.
-     *
-     * @param array<string, mixed> $data
-     */
-    private function jsonResponseWithRateLimitHeaders(
-        array $data,
-        RateLimitResult $rateLimitResult,
-        int $statusCode = 200,
-    ): JsonResponse {
-        $response = new JsonResponse($data, $statusCode);
-
-        foreach ($rateLimitResult->getHeaders() as $name => $value) {
-            $response = $response->withAddedHeader($name, $value);
-        }
-
-        return $response;
     }
 
     /**
