+++ Classes/Controller/TemplateController.php
@@ -12,13 +12,11 @@ namespace Netresearch\T3Cowriter\Controller;
 use Netresearch\NrLlm\Domain\Model\Task;
 use Netresearch\NrLlm\Domain\Repository\TaskRepository;
 use Netresearch\T3Cowriter\Service\RateLimiterInterface;
-use Netresearch\T3Cowriter\Service\RateLimitResult;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 use Throwable;
 use TYPO3\CMS\Core\Context\Context;
-use TYPO3\CMS\Core\Http\JsonResponse;
 
 /**
  * AJAX controller for prompt template management.
@@ -29,6 +27,8 @@ use TYPO3\CMS\Core\Http\JsonResponse;
  */
 final readonly class TemplateController
 {
+    use RateLimitedControllerTrait;
+
     public function __construct(
         private TaskRepository $templateRepository,
         private RateLimiterInterface $rateLimiter,
@@ -78,38 +78,5 @@ final readonly class TemplateController
                 500,
             );
         }
-    }
-
-    /**
-     * Create JSON response with rate limit headers.
-     *
-     * @param array<string, mixed> $data
-     */
-    private function jsonResponseWithRateLimitHeaders(
-        array $data,
-        RateLimitResult $rateLimitResult,
-        int $statusCode = 200,
-    ): JsonResponse {
-        $response = new JsonResponse($data, $statusCode);
-
-        foreach ($rateLimitResult->getHeaders() as $name => $value) {
-            $response = $response->withAddedHeader($name, $value);
-        }
-
-        return $response;
-    }
-
-    private function rateLimitedResponse(RateLimitResult $result): JsonResponse
-    {
-        $response = new JsonResponse(
-            ['success' => false, 'error' => 'Rate limit exceeded. Please try again later.'],
-            429,
-        );
-
-        foreach ($result->getHeaders() as $name => $value) {
-            $response = $response->withAddedHeader($name, $value);
-        }
-
-        return $response->withAddedHeader('Retry-After', (string) $result->getRetryAfter());
     }
 }
