+++ Classes/Controller/ToolController.php
@@ -9,20 +9,17 @@ declare(strict_types=1);
 
 namespace Netresearch\T3Cowriter\Controller;
 
-use JsonException;
 use Netresearch\NrLlm\Domain\ValueObject\ToolSpec;
 use Netresearch\NrLlm\Service\LlmServiceManagerInterface;
 use Netresearch\NrLlm\Service\Option\ToolOptions;
 use Netresearch\T3Cowriter\Domain\DTO\ToolRequest;
 use Netresearch\T3Cowriter\Service\RateLimiterInterface;
-use Netresearch\T3Cowriter\Service\RateLimitResult;
 use Netresearch\T3Cowriter\Tools\ContentQueryTool;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 use Throwable;
 use TYPO3\CMS\Core\Context\Context;
-use TYPO3\CMS\Core\Http\JsonResponse;
 
 /**
  * AJAX controller for LLM tool calling via nr-llm.
@@ -34,6 +31,8 @@ use TYPO3\CMS\Core\Http\JsonResponse;
  */
 final readonly class ToolController
 {
+    use RateLimitedControllerTrait;
+
     public function __construct(
         private LlmServiceManagerInterface $llmServiceManager,
         private RateLimiterInterface $rateLimiter,
@@ -139,65 +138,5 @@ final readonly class ToolController
         }
 
         return $resolved !== [] ? $resolved : array_values($allTools);
-    }
-
-    /**
-     * Parse JSON body from request, returning null on failure.
-     *
-     * @return array<string, mixed>|null
-     */
-    private function parseJsonBody(ServerRequestInterface $request): ?array
-    {
-        $rawBody = (string) $request->getBody();
-        if ($rawBody === '') {
-            return [];
-        }
-
-        try {
-            /** @var mixed $decoded */
-            $decoded = json_decode($rawBody, true, 512, JSON_THROW_ON_ERROR);
-        } catch (JsonException) {
-            return null;
-        }
-
-        if (!is_array($decoded)) {
-            return null;
-        }
-
-        /** @var array<string, mixed> $decoded */
-        return $decoded;
-    }
-
-    /**
-     * Create JSON response with rate limit headers.
-     *
-     * @param array<string, mixed> $data
-     */
-    private function jsonResponseWithRateLimitHeaders(
-        array $data,
-        RateLimitResult $rateLimitResult,
-        int $statusCode = 200,
-    ): JsonResponse {
-        $response = new JsonResponse($data, $statusCode);
-
-        foreach ($rateLimitResult->getHeaders() as $name => $value) {
-            $response = $response->withAddedHeader($name, $value);
-        }
-
-        return $response;
-    }
-
-    private function rateLimitedResponse(RateLimitResult $result): JsonResponse
-    {
-        $response = new JsonResponse(
-            ['success' => false, 'error' => 'Rate limit exceeded. Please try again later.'],
-            429,
-        );
-
-        foreach ($result->getHeaders() as $name => $value) {
-            $response = $response->withAddedHeader($name, $value);
-        }
-
-        return $response->withAddedHeader('Retry-After', (string) $result->getRetryAfter());
     }
 }
